requirements.txt
# Audio
pyaudio>=0.2.13
numpy>=1.24.0

# Speech-to-Text
openai-whisper>=20231117
# or use whisper-cpp for faster inference

# Text-to-Speech (ElevenLabs streaming)
websockets>=12.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0

# Deepgram streaming STT (enabled for low latency)
deepgram-sdk>=3.0.0

# Tier 1 Intent Routing (FAISS embedding search)
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# Tier 2 Local LLM Routing (MLX on Apple Silicon)
mlx>=0.29.0
mlx-lm>=0.22.0
